refactor out addition, multiplication, and pairing into separate modules
